[MIG] Migration to 12.0: product_history_for_cpo
