--- intercoop_addons/product_history_for_cpo/i18n/fr.po
@@ -1,34 +1,0 @@
-# Translation of Odoo Server.
-# This file contains the translation of the following modules:
-#	* product_history_for_cpo
-#
-msgid ""
-msgstr ""
-"Project-Id-Version: Odoo Server 9.0c\n"
-"Report-Msgid-Bugs-To: \n"
-"POT-Creation-Date: 2016-03-02 10:03+0000\n"
-"PO-Revision-Date: 2016-03-02 10:03+0000\n"
-"Last-Translator: <>\n"
-"Language-Team: \n"
-"MIME-Version: 1.0\n"
-"Content-Type: text/plain; charset=UTF-8\n"
-"Content-Transfer-Encoding: \n"
-"Plural-Forms: \n"
-
-#. module: product_history_for_cpo
-#: model:ir.model,name:product_history_for_cpo.model_computed_purchase_order_line
-msgid "Computed Purchase Order Line"
-msgstr "Ligne de calcul de commande d'achat"
-
-#. module: product_history_for_cpo
-#: model:ir.model.fields,field_description:product_history_for_cpo.field_computed_purchase_order_line_displayed_product_history_ids
-msgid "Product History"
-msgstr "Historique Produit"
-
-#. module: product_history_for_cpo
-#: model:ir.actions.act_window,name:product_history_for_cpo.action_view_history
-#: model:ir.model,name:product_history_for_cpo.model_view_product_history
-#: model:ir.ui.view,arch_db:product_history_for_cpo.view_computed_purchase_order_form
-msgid "View Product History"
-msgstr "Historique Produit"
-
