--- intercoop_addons/product_history_for_cpo/i18n/product_history_for_cpo.pot
@@ -1,47 +1,0 @@
-# Translation of Odoo Server.
-# This file contains the translation of the following modules:
-#	* product_history_for_cpo
-#
-msgid ""
-msgstr ""
-"Project-Id-Version: Odoo Server 9.0c\n"
-"Report-Msgid-Bugs-To: \n"
-"Last-Translator: <>\n"
-"Language-Team: \n"
-"MIME-Version: 1.0\n"
-"Content-Type: text/plain; charset=UTF-8\n"
-"Content-Transfer-Encoding: \n"
-"Plural-Forms: \n"
-
-#. module: product_history_for_cpo
-#: model:ir.model,name:product_history_for_cpo.model_computed_purchase_order_line
-msgid "Computed Purchase Order Line"
-msgstr ""
-
-#. module: product_history_for_cpo
-#: model:ir.model.fields,field_description:product_history_for_cpo.field_computed_purchase_order_line_displayed_product_history_ids
-#: model:ir.ui.view,arch_db:product_history_for_cpo.view_product_history_tree
-msgid "Product History"
-msgstr ""
-
-#. module: product_history_for_cpo
-#: model:ir.ui.view,arch_db:product_history_for_cpo.view_product_history_tree
-msgid "This history line is active. Click this button to ignore it."
-msgstr ""
-
-#. module: product_history_for_cpo
-#: model:ir.ui.view,arch_db:product_history_for_cpo.view_product_history_tree
-msgid "This history line is active. Click this button to reactivate it."
-msgstr ""
-
-#. module: product_history_for_cpo
-#: model:ir.actions.act_window,name:product_history_for_cpo.action_view_history
-#: model:ir.ui.view,arch_db:product_history_for_cpo.view_computed_purchase_order_form
-msgid "View Product History"
-msgstr ""
-
-#. module: product_history_for_cpo
-#: model:ir.model,name:product_history_for_cpo.model_product_history
-msgid "product.history"
-msgstr ""
-
