+++ product_history_for_cpo/views/product_history_view.xml
@@ -1,16 +1,12 @@
 <?xml version="1.0" encoding="utf-8"?>
-<!-- ********************************************************************** -->
-<!-- See __openerp__.py file for Copyright and Licence Informations.         -->
-<!-- ********************************************************************** -->
-<openerp>
+<odoo>
     <data>
-<!-- View : computed.purchase.order -->
-        <record id="view_computed_purchase_order_form" model="ir.ui.view">
+        <record id="view_computed_purchase_order_inherit_form" model="ir.ui.view">
             <field name="inherit_id" ref="purchase_compute_order.view_computed_purchase_order_form"/>
             <field name="model">computed.purchase.order</field>
             <field name="arch" type="xml">
                 <field name="product_code_inv" position="before">
-                    <button name="view_history" string="View Product History" type="object" icon="STOCK_FIND"/>
+                    <button name="view_history" string="View Product History" type="object" icon="fa-list fa-2x"/>
                 </field>
             </field>
         </record>
@@ -30,10 +26,21 @@
                     <field name="outgoing_qty"/>
                     <field name="virtual_qty"/>
                     <field name="ignored" invisible="1"/>
-                    <button name="ignore_line_cpo" attrs="{'invisible': [('ignored','=',True)]}" type="object" icon="STOCK_APPLY" help="This history line is active. Click this button to ignore it."/>
-                    <button name="unignore_line_cpo" attrs="{'invisible': [('ignored','=',False)]}" type="object" icon="STOCK_STOP" help="This history line is active. Click this button to reactivate it."/>
+                    <button name="ignore_line_cpo" attrs="{'invisible': [('ignored','=',True)]}" type="object" icon="fa-plus-circle fa-2x" help="This history line is active. Click this button to ignore it."/>
+                    <button name="unignore_line_cpo" attrs="{'invisible': [('ignored','=',False)]}" type="object" icon="fa-times-circle fa-2x" help="This history line is active. Click this button to reactivate it."/>
                 </tree>
             </field>
         </record>
+
+        <record id="action_view_history" model="ir.actions.act_window">
+            <field name="name">View Product History</field>
+            <field name="type">ir.actions.act_window</field>
+            <field name="res_model">product.history</field>
+            <field name="view_type">form</field>
+            <field name="view_mode">tree</field>
+            <field name="view_id" ref="product_history_for_cpo.view_product_history_tree"/>
+            <field name="target">new</field>
+            <field name="domain">[]</field>
+        </record>
     </data>
-</openerp>
+</odoo>
